requirements.txt
Backtesting==0.3.3
bokeh==3.6.2
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.1
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyTelegramBotAPI==4.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
ta==0.11.0
tornado==6.4.2
tzdata==2024.2
urllib3==2.3.0
xyzservices==2025.1.0
